🏗 Modularizando o texto com glitch
