+++ src/components/sobre/sobre.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import {
-  Box,
   Heading,
   PseudoBox,
   Text,
@@ -8,22 +7,8 @@ import {
   from '@chakra-ui/core';
 import { Global } from '@emotion/core';
 
-import grafo from '../images/grafo.svg';
-
-let e;
-
-const startGlitching = () => {
-  e = setInterval(() => {
-    document.documentElement.style.setProperty('--clip-path-1', `inset(${Math.random() * 100}% 0% ${Math.random() * 100}% 0%)`);
-    document.documentElement.style.setProperty('--clip-path-2', `inset(${Math.random() * 100}% 0% ${Math.random() * 100}% 0%)`);
-  }, 150);
-};
-
-const stopGlitching = () => {
-  clearInterval(e);
-  document.documentElement.style.setProperty('--clip-path-1', 'inset(100% 0% 0% 0%)');
-  document.documentElement.style.setProperty('--clip-path-2', 'inset(100% 0% 0% 0%)');
-};
+import GlitchingText from './glitchingText';
+import grafo from '../../images/grafo.svg';
 
 const Sobre = () => (
   <PseudoBox
@@ -73,34 +58,7 @@ const Sobre = () => (
       apresentar tópicos que lhe interessam.
     </Text>
     <Heading size="xl" my={5} textAlign="center">
-      <PseudoBox
-        display="inline"
-        position="relative"
-        zIndex="2"
-        _before={{
-          bg: 'cyan.500',
-          clipPath: 'var(--clip-path-1)',
-          position: 'absolute',
-          bottom: 0,
-          content: '"SAC Virtual?"',
-          left: '2px',
-          textShadow: '0 -2px #f0f',
-          width: 'calc(100% + 2px)',
-        }}
-        _after={{
-          bg: 'cyan.500',
-          clipPath: 'var(--clip-path-2)',
-          position: 'absolute',
-          bottom: 0,
-          content: '"SAC Virtual?"',
-          left: '-3px',
-          textShadow: '0 3px #0ff',
-        }}
-        onMouseEnter={startGlitching}
-        onMouseLeave={stopGlitching}
-      >
-        SAC Virtual?
-      </PseudoBox>
+      <GlitchingText>SAC Virtual?</GlitchingText>
     </Heading>
     <Text maxW="90ch" mt={2} mb={10} mx="auto">
       Em decorrência da pandemia da COVID-19, a XV SAC será virtual, para você poder
