+++ src/pages/index.jsx
@@ -3,7 +3,7 @@ import { ThemeProvider, CSSReset } from '@chakra-ui/core';
 
 import SEO from '../components/seo';
 import Header from '../components/header';
-import Sobre from '../components/sobre';
+import Sobre from '../components/sobre/sobre';
 import Timeline from '../components/timeline';
 import Parceiros from '../components/parceiros';
 import Footer from '../components/footer';
